Polishing to a Demo State #1
- Rotational Velocity can be edited for projectiles that use it, and
slowed in Time Bubble
- Improved Item Pickup Mechanics so it is simplified and size test is
performed on Pickup Action being pressed
- Fixed DWeapon to not show ammo for Gravity Gun, and turn off when
dropped
- Switched Health and Stamina variables to float so that damage
calculations can be more complex (this was proving difficult with either
ints or bytes)
- Above allowed for Damage and throw speed values to change based on
player and item size ratio
-DSWeapons now have scale limit

+++ Source/BrensFullGame/BCharacter.h
@@ -16,19 +16,19 @@ public:
 
 	// Variable Controlling Current health
 	UPROPERTY(EditAnywhere, Category = "Main", BlueprintReadWrite)
-	uint8 CurrentHealth;
+	float CurrentHealth;
 
 	// Variable Controlling Max health
 	UPROPERTY(EditAnywhere, Category = "Main", BlueprintReadWrite)
-	uint8 MaxtHealth;
+	float MaxtHealth;
 
 	// Variable Controlling Current Stamina
 	UPROPERTY(EditAnywhere, Category = "Main", BlueprintReadWrite)
-	uint8 CurrentStamina;
+	float CurrentStamina;
 
 	// Variable Controlling Max Stamina
 	UPROPERTY(EditAnywhere, Category = "Main", BlueprintReadWrite)
-	uint8 MaxStamina;
+	float MaxStamina;
 
 	// Bool controlling whether Item is held or not
 	UPROPERTY(EditAnywhere, Category = "Main", BlueprintReadWrite)
